Add files via upload
specs sheet and board revisions 1A (has !CS errorenously connected to SN which is a sysIO that, as it turned out, cannot be made general purpose IO while simultaneously using the SPI EFB... sad because the PCBs ever manufactured are 1A), can be fixed by cutting the pads in half and soldering some wires. Sorry!) and 1B (problem fixed).
The test pad is there to enable JTAG (needs to be pulled high, otherwise JTAG is repurposed for the 245 interface.